refactor: 全项目 5 维审计 14 条修复 + CI 覆盖 api/cmd + api 文档双语化
5 维（设计/测试/权限/冗余/注释）审计的可执行清单全修，另补 1 处契约漂移：

- 契约/行为
  - validator stream 校验失败码退化为 Unknown -> 统一走 errs.InvalidArgument
  - /api/ws_info 与 /ws 经核实为免登录探活端点，doc/access_control.md 标注对齐实现
  - namespace.Favorite 落 RequireNamespaceAccessByID 后同步 doc 契约行
- 测试缺口：CopyToPod 的 GetByID 错误分支、changelog 的 DeployDailyCounts/dayKey 补测
- 去过度抽象：k8sutil.PathSpec、wssender/nsq 的 nsqProducer|nsqConsumer
  三处「唯一实现 + 零 fake」接口收敛为具体类型
- 清理：冗余类型转换与未使用参数；RequireFileAccess 注释补 CopyToPod 消费方
- 注释：导出符号补注释、纯英文散文注释转中文、考古注释与过程引用清除
- 命名：GrpcServerImp -> GrpcServerImpl（接口名漏 l），mock 同步重生成
- CI：新增 make test_all（根模块非 internal 包 + 独立模块 api/），workflow 改调它
- 文档：api/README 双语化（新增英文版，原文件更名 README_zh-CN.md），方法数订正
- 构建：Dockerfile 拆出 go mod download 前置层 + BuildKit go-build 缓存挂载，
  gitBranch 在 detached HEAD 下回退 tag 名；build_image workflow 改用
  docker/build-push-action 并接 GHA 缓存，修复 tag 触发时版本接口暴露 "HEAD"

验证：make test_all EXIT=0（0 FAIL / 覆盖 97.4%）、make lint 0 issues、
make sec 0 命中、gofmt 净、根模块与 api 模块 build+vet 全过。

+++ internal/util/date/date.go
@@ -91,11 +91,11 @@ func HumanDuration(d time.Duration) string {
 		return fmt.Sprintf("%d天", hours/24)
 	}
 	if hours < 24*365*8 {
-		dy := int(hours/24) % 365
+		dy := hours / 24 % 365
 		if dy == 0 {
 			return fmt.Sprintf("%d年", hours/24/365)
 		}
 		return fmt.Sprintf("%d年%d天", hours/24/365, dy)
 	}
-	return fmt.Sprintf("%d年", int(hours/24/365))
+	return fmt.Sprintf("%d年", hours/24/365)
 }
